+++ src/table.ts
@@ -5,8 +5,5 @@ import type { Property } from "csstype";
 /** Default CSS `caption-side` property value. */
 export const captionSide: Property.CaptionSide = "top";
 
-/** Reverse of default CSS `caption-side` property value. */
-export const reverseCaptionSide: Property.CaptionSide = "bottom";
-
 /** Default CSS `empty-cells` property value. */
 export const emptyCells: Property.EmptyCells = "show";